short -> int

+++ TombEditor/Command.cs
@@ -97,7 +97,7 @@ namespace TombEditor
             }
         }
 
-        private static void GenericDirectionalControlCommand(CommandArgs args, BlockVertical surface, short increment, bool smooth, bool oppositeDiagonal)
+        private static void GenericDirectionalControlCommand(CommandArgs args, BlockVertical surface, int increment, bool smooth, bool oppositeDiagonal)
         {
             if (args.Editor.LastSelection == LastSelectionType.Block && args.Editor.Mode == EditorMode.Geometry && args.Editor.SelectedSectors.Valid)
             {
@@ -414,17 +414,17 @@ namespace TombEditor
 
             AddCommand("RaiseQA4Click", "Raise selected floor or item (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)(args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("LowerQA1Click", "Lower selected floor or item (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)-args.Editor.IncrementReference, false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, -args.Editor.IncrementReference, false, false);
             });
 
             AddCommand("LowerQA4Click", "Lower selected floor or item (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)(-args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, -args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("RaiseWS1Click", "Raise selected ceiling (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -434,17 +434,17 @@ namespace TombEditor
 
             AddCommand("RaiseWS4Click", "Raise selected ceiling (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)(args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("LowerWS1Click", "Lower selected ceiling (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)-args.Editor.IncrementReference, false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, -args.Editor.IncrementReference, false, false);
             });
 
             AddCommand("LowerWS4Click", "Lower selected ceiling (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)(-args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, -args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("RaiseED1Click", "Raise selected floor subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -454,17 +454,17 @@ namespace TombEditor
 
             AddCommand("RaiseED4Click", "Raise selected floor subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, (short)(args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("LowerED1Click", "Lower selected floor subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, (short)-args.Editor.IncrementReference, false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, -args.Editor.IncrementReference, false, false);
             });
 
             AddCommand("LowerED4Click", "Lower selected floor subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, (short)(-args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, -args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("RaiseRF1Click", "Raise selected ceiling subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -474,17 +474,17 @@ namespace TombEditor
 
             AddCommand("RaiseRF4Click", "Raise selected ceiling subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, (short)(args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("LowerRF1Click", "Lower selected ceiling subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, (short)-args.Editor.IncrementReference, false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, -args.Editor.IncrementReference, false, false);
             });
 
             AddCommand("LowerRF4Click", "Lower selected ceiling subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, (short)(-args.Editor.IncrementReference * 4), false, false);
+                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, -args.Editor.IncrementReference * 4, false, false);
             });
 
             AddCommand("RaiseQA1ClickSmooth", "Smoothly raise selected floor (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -494,17 +494,17 @@ namespace TombEditor
 
             AddCommand("RaiseQA4ClickSmooth", "Smoothly raise selected floor (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)(args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("LowerQA1ClickSmooth", "Smoothly lower selected floor (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)-args.Editor.IncrementReference, true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, -args.Editor.IncrementReference, true, false);
             });
 
             AddCommand("LowerQA4ClickSmooth", "Smoothly lower selected floor (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)(-args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, -args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("RaiseWS1ClickSmooth", "Smoothly raise selected ceiling (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -514,17 +514,17 @@ namespace TombEditor
 
             AddCommand("RaiseWS4ClickSmooth", "Smoothly raise selected ceiling (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)(args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("LowerWS1ClickSmooth", "Smoothly lower selected ceiling (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)-args.Editor.IncrementReference, true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, -args.Editor.IncrementReference, true, false);
             });
 
             AddCommand("LowerWS4ClickSmooth", "Smoothly lower selected ceiling (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)(-args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, -args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("RaiseED1ClickSmooth", "Smoothly raise selected floor subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -534,17 +534,17 @@ namespace TombEditor
 
             AddCommand("RaiseED4ClickSmooth", "Smoothly raise selected floor subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, (short)(args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("LowerED1ClickSmooth", "Smoothly lower selected floor subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, (short)-args.Editor.IncrementReference, true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, -args.Editor.IncrementReference, true, false);
             });
 
             AddCommand("LowerED4ClickSmooth", "Smoothly lower selected floor subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, (short)(-args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.FloorSubdivision2, -args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("RaiseRF1ClickSmooth", "Smoothly raise selected ceiling subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -554,17 +554,17 @@ namespace TombEditor
 
             AddCommand("RaiseRF4ClickSmooth", "Smoothly raise selected ceiling subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, (short)(args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("LowerRF1ClickSmooth", "Smoothly lower selected ceiling subdivision (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, (short)-args.Editor.IncrementReference, true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, -args.Editor.IncrementReference, true, false);
             });
 
             AddCommand("LowerRF4ClickSmooth", "Smoothly lower selected ceiling subdivision (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, (short)(-args.Editor.IncrementReference * 4), true, false);
+                GenericDirectionalControlCommand(args, BlockVertical.CeilingSubdivision2, -args.Editor.IncrementReference * 4, true, false);
             });
 
             AddCommand("RaiseYH1Click", "Raise selected floor diagonal step (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -574,17 +574,17 @@ namespace TombEditor
 
             AddCommand("RaiseYH4Click", "Raise selected floor diagonal step (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)(args.Editor.IncrementReference * 4), false, true);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, args.Editor.IncrementReference * 4, false, true);
             });
 
             AddCommand("LowerYH1Click", "Lower selected floor diagonal step (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)-args.Editor.IncrementReference, false, true);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, -args.Editor.IncrementReference, false, true);
             });
 
             AddCommand("LowerYH4Click", "Lower selected floor diagonal step (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Floor, (short)(-args.Editor.IncrementReference * 4), false, true);
+                GenericDirectionalControlCommand(args, BlockVertical.Floor, -args.Editor.IncrementReference * 4, false, true);
             });
 
             AddCommand("RaiseUJ1Click", "Raise selected ceiling diagonal step (1 click)", CommandType.Geometry, delegate (CommandArgs args)
@@ -594,17 +594,17 @@ namespace TombEditor
 
             AddCommand("RaiseUJ4Click", "Raise selected ceiling diagonal step (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)(args.Editor.IncrementReference * 4), false, true);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, args.Editor.IncrementReference * 4, false, true);
             });
 
             AddCommand("LowerUJ1Click", "Lower selected ceiling diagonal step (1 click)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)-args.Editor.IncrementReference, false, true);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, -args.Editor.IncrementReference, false, true);
             });
 
             AddCommand("LowerUJ4Click", "Lower selected ceiling diagonal step (4 clicks)", CommandType.Geometry, delegate (CommandArgs args)
             {
-                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, (short)(-args.Editor.IncrementReference * 4), false, true);
+                GenericDirectionalControlCommand(args, BlockVertical.Ceiling, -args.Editor.IncrementReference * 4, false, true);
             });
 
             AddCommand("RotateObject5", "Rotate object (5 degrees)", CommandType.Objects, delegate (CommandArgs args)
