+++ TombLib/TombLib/LevelData/Compilers/TombEngine/LevelCompilerTombEngine.cs
@@ -389,7 +389,7 @@ namespace TombLib.LevelData.Compilers.TombEngine
             }
         }
 
-        private short GetMostDownFloor(Room room, int x, int z)
+        private int GetMostDownFloor(Room room, int x, int z)
         {
             FindBottomFloor(ref room, ref x, ref z);
             return _tempRooms[room].AuxSectors[x, z].LowestFloor;
